fix(cache): extract flag cache store
Keep DefaultFlagRepository under the Detekt constructor threshold while preserving the Room-backed flags cache.

Validation: ./scripts/docker-dev.sh ./gradlew detektAll :core:data:testDebugUnitTest
