+++ core/data/src/test/kotlin/fr/forumhfr/redface2/core/data/flags/DefaultFlagRepositoryTest.kt
@@ -625,8 +625,11 @@ class DefaultFlagRepositoryTest {
         apiClient = apiClient,
         forumRepository = forumRepository,
         authRepository = authRepository,
-        flagDao = flagDao,
-        clock = clock,
+        flagCacheStore = FlagCacheStore(
+            flagDao = flagDao,
+            clock = clock,
+            ioDispatcher = UnconfinedTestDispatcher(),
+        ),
         json = json,
         ioDispatcher = UnconfinedTestDispatcher(),
     )
